Moved the get_mission_and_observer_and_target method from the Product to the Collection class.
